+++ actix/tests/test_fut.rs
@@ -45,6 +45,7 @@ fn test_fut_timeout() {
 }
 
 struct MyStreamActor {
+    counter: usize,
     timeout: Arc<AtomicBool>,
 }
 
@@ -60,13 +61,11 @@ impl Actor for MyStreamActor {
             .timeout(Duration::new(0, 1000))
             .then(|res, act, _| {
                 // Additional waiting time to test `then` call as well
-                Box::pin(
-                    async move {
-                        sleep(Duration::from_millis(500)).await;
-                        res
-                    }
-                    .into_actor(act),
-                )
+                async move {
+                    sleep(Duration::from_millis(500)).await;
+                    res
+                }
+                .into_actor(act)
             })
             .map(|e, act, _| {
                 if let Err(()) = e {
@@ -79,6 +78,96 @@ impl Actor for MyStreamActor {
     }
 }
 
+struct TakeWhileMsg(usize);
+
+impl Message for TakeWhileMsg {
+    type Result = ();
+}
+
+impl Handler<TakeWhileMsg> for MyStreamActor {
+    type Result = ResponseActFuture<Self, ()>;
+
+    fn handle(&mut self, msg: TakeWhileMsg, _: &mut Context<Self>) -> Self::Result {
+        let num = msg.0;
+        Box::pin(
+            futures_util::stream::repeat(num)
+                .into_actor(self)
+                .take_while(move |n, act, ctx| {
+                    ctx.spawn(
+                        async {
+                            actix_rt::task::yield_now().await;
+                        }
+                        .into_actor(act),
+                    );
+                    assert_eq!(*n, num);
+                    assert!(act.counter < 10);
+                    act.counter += 1;
+                    futures_util::future::ready(act.counter < 10)
+                })
+                .finish(),
+        )
+    }
+}
+
+struct SkipWhileMsg(usize);
+
+impl Message for SkipWhileMsg {
+    type Result = ();
+}
+
+impl Handler<SkipWhileMsg> for MyStreamActor {
+    type Result = ResponseActFuture<Self, ()>;
+
+    fn handle(&mut self, msg: SkipWhileMsg, _: &mut Context<Self>) -> Self::Result {
+        let num = msg.0;
+        Box::pin(
+            futures_util::stream::repeat(num)
+                .into_actor(self)
+                .take_while(|_, act, _| {
+                    let cond = act.counter < 10;
+                    act.counter += 1;
+                    futures_util::future::ready(cond)
+                })
+                .skip_while(move |n, act, ctx| {
+                    let fut = async {
+                        actix_rt::task::yield_now().await;
+                    }
+                    .into_actor(act);
+                    ctx.spawn(fut);
+                    assert_eq!(*n, num);
+                    act.counter -= 1;
+                    futures_util::future::ready(act.counter > 0)
+                })
+                .finish(),
+        )
+    }
+}
+
+struct CollectMsg(usize);
+
+impl Message for CollectMsg {
+    type Result = Vec<usize>;
+}
+
+impl Handler<CollectMsg> for MyStreamActor {
+    type Result = ResponseActFuture<Self, Vec<usize>>;
+
+    fn handle(&mut self, msg: CollectMsg, _: &mut Context<Self>) -> Self::Result {
+        let num = msg.0;
+        Box::pin(
+            futures_util::stream::repeat(num)
+                .into_actor(self)
+                .take_while(|_, act, _| {
+                    let cond = act.counter < 5;
+                    act.counter += 1;
+                    futures_util::future::ready(cond)
+                })
+                .map(|_, act, _| act.counter)
+                .collect(),
+        )
+    }
+}
+
 #[test]
 fn test_stream_timeout() {
     let timeout = Arc::new(AtomicBool::new(false));
@@ -86,9 +175,51 @@ fn test_stream_timeout() {
 
     let sys = System::new();
     sys.block_on(async {
-        let _addr = MyStreamActor { timeout: timeout2 }.start();
+        let _addr = MyStreamActor {
+            timeout: timeout2,
+            counter: 0,
+        }
+        .start();
     });
     sys.run().unwrap();
 
     assert!(timeout.load(Ordering::Relaxed), "Not timeout");
+}
+
+#[test]
+fn test_stream_take_while() {
+    System::new().block_on(async {
+        let addr = MyStreamActor {
+            timeout: Arc::new(AtomicBool::new(false)),
+            counter: 0,
+        }
+        .start();
+        addr.send(TakeWhileMsg(5)).await.unwrap();
+    })
+}
+
+#[test]
+fn test_stream_skip_while() {
+    System::new().block_on(async {
+        let addr = MyStreamActor {
+            timeout: Arc::new(AtomicBool::new(false)),
+            counter: 0,
+        }
+        .start();
+        addr.send(SkipWhileMsg(5)).await.unwrap();
+    })
+}
+
+#[test]
+fn test_stream_collect() {
+    System::new().block_on(async {
+        let addr = MyStreamActor {
+            timeout: Arc::new(AtomicBool::new(false)),
+            counter: 0,
+        }
+        .start();
+        let res = addr.send(CollectMsg(3)).await.unwrap();
+
+        assert_eq!(res, vec![1, 2, 3, 4, 5]);
+    })
 }